Records the number of sources per source origin on DB initialization
We currently have a limit on the number of sources per source origin. To
better understand the impact on disk usage of this limit, we want to
understand the distribution that follows the number of sources per
source origin.

In this CL we add four histograms that correspond to the number of
sources per source origin for the nth ranked source origin.

- Conversions.SourcesPerSourceOrigin.1st
- Conversions.SourcesPerSourceOrigin.3rd
- Conversions.SourcesPerSourceOrigin.7th
- Conversions.SourcesPerSourceOrigin.20th

The number of selected positions and their ranks is somewhat arbitrary
but they were picked to have sufficient samples from the distribution
while limiting the number of histograms added.

Change-Id: I06008628c443f93ec4a752ad97a5182ecdd9b7a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4864573
Commit-Queue: Anthony Garant <anthonygarant@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198997}