privacy-hub: Make the microphone toggle functional
CL:4813879 changed the view of the microphone section in the Privacy
controls page. But the access control toggle was not connected to the
pref. This CL connects the toggle to the underlying pref.

This CL also removes the list of microphones from the Privacy controls
page.

Bug: b:300915258
Change-Id: I6041a9d325dab63feb40afe0d1f4195b8d5887d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4873498
Reviewed-by: Wes Okuhara <wesokuhara@google.com>
Commit-Queue: Md Shahadat Hossain Shahin <shahinmd@google.com>
Cr-Commit-Position: refs/heads/main@{#1199058}