[ForceSignin] Move the reauth flow to the profile picker
Add a step to the ProfilePickerFlowController to be able to reauth.
Uses the newly created ProfilePickerReauthProvider to handle the
reauth within the Gaia reauth page (/AccountChooser will show the
"Verify it's you" page).

Follow up:
- React to a user reauth-ing using a different account, but an
account that was already signed in (it is currently not detected).
- React the success == false where the reauth either failed or was
done on a different account.
- Add the back button and it's equivalent reactions to the reauth
page.

Bug: 1478217
Change-Id: I68d3550b2d62c4c7c509f049937fded83d20201a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4856638
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Ryan Sultanem <rsult@google.com>
Cr-Commit-Position: refs/heads/main@{#1199060}