cleanup: feedv2: remove java-side cached flag for InterestFeedV2
The flag is enabled by default and not expected to be disabled in the
field at any point in the future. The trial has expired, so the code is
always reading the defaults.

The InterestFeedV2 feature is used to fetch the referrer_url param.
Hopefully this does not need to be configurable: only one value is used
in the field right now.

The InterestFeedV2 is used to fetch more performance-related knobs on
the native side. Assuming we want to keep them. This leads to keeping
the native-side base::Feature(kInterestFeedV2).

Removing the (native) kInterestFeedContentSuggestions flag because it
is overridden by kInterestFeedV2. Configuring it on server-side would be
risky.

The string constants (INTEREST_FEED_V2 and
INTEREST_FEED_CONTENT_SUGGESTIONS) are referenced in @EnableFeatures in
tests. They override nothing, so it should be safe to remove them,
right? Right? Removing.

Added batching for FeedSurfaceScopeDependencyProviderImplTest because
the test looked simple enough to not depend on global state. And because
presubmit was feeding me with a warning, of course.

Bug: 1069183
Change-Id: Ic278813b7d5d3de10c7bdffe645d90e30cda9b40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4866287
Reviewed-by: Justin DeWitt <dewittj@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Commit-Queue: Egor Pasko <pasko@chromium.org>
Reviewed-by: Dan H <harringtond@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1199072}