[webauthn] Fix chromeos credential availability for google.com
crrev.com/c/4822929 updated Chrome to filter for google.com that are not
used for authentication. The logic assumed that the platform
authenticator can always enumerate credentials, and would clear the
`HasPlatformAuthenticatorCredential` flag if that was not the case.
However, ChromeOS cannot enumerate credentials, so this broke the
platform authenticator and the power button.

This CL adds a restriction to only filter credentials if platform
credentials were enumerated, which should fix the bug.

Bug: 1480732
Change-Id: Ib0bb2b7db90719b2e886a28ee6aab774badb6a65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4874424
Auto-Submit: Nina Satragno <nsatragno@chromium.org>
Reviewed-by: Adam Langley <agl@chromium.org>
Commit-Queue: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199077}