Changes hard crashes to DCHECKs for DevTools histograms
DevTools histogram reports would trigger hard crashes when out-of-range values were transmitted to the embedder from the front-end side. This doesn't really make a lot of sense; while it is desirable to have some insight to failing or improperly-ranged histograms, having a hard crash that occurs in all cases doesn't make a lot of sense because the histogram isn't something that can necessarily be controlled by the customer.

This change handles the count- and enumerated histogram types with the
following behavior:

 - for both, we will DCHECK if compiled with DCHECKs enabled
 - if DCHECKs are off:
   - for count-type, we clamp to the client-specified acceptable range
   - for enumerated-type, we skip recording the histogram if it's out of the acceptable range

Bug: 1482914
Change-Id: I8e96602cad4e82b69b4c54295e9fadc860e196cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4874788
Commit-Queue: Robert Paveza <Rob.Paveza@microsoft.com>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199087}