Require device pin/password to retrieve saved Chrome passwords in automotive
The biometric reauth feature uses DeviceAuthenticatorBridge#authenticate
API, which will display the pin/password prompt if that's how the device
is secured, so we can simply enable this feature for Android automotive
to require pin/password for retrieving passwords.

Note that the biometric reauth feature was not rolled out on Android
because it seemed to cause users to avoid filling passwords altogether
and there wasn't a toggle provided for users to turn this feature off.
After UPM (and settings migration to GMS Core), the feature will be
rolled out on Android in general. When that happens, we will need to
disable that setting on automotive: crbug/1479776.

Bug: 1475341
Change-Id: I3790fe177917a4fc19af5d7e68d9d7b017e1da4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4844948
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Geoff Huang <geoffhuang@google.com>
Cr-Commit-Position: refs/heads/main@{#1199134}