Move user agent from CriticalPersistedTabData to Tab
This is part of CriticalPersistedTabData deprecation to reverse
the usage of the UserData pattern. Persistence parts of
CriticalPersistedTabData are filled in with placeholder values
for user agent. This is safe to do because the experiment is turned
off.

Bug: b/301154780
Change-Id: Id159ff37722fe842044973e52578ffcdefd2857b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4875130
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: David Maunder <davidjm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199138}