[webauthn] Add platform authenticator button if it can't enumerate creds
With the new passkey get assertion UI, platform authenticator
credentials are listed in the mechanism selection screen, so we don't
show a dedicated button to dispatch to the platform authenticator.
Unfortunately, the ChromeOS platform authenticator can report the
availability of credentials but not enumerate them, so there is no way
to trigger it with the new UI.

This CL fixes this bug by bringing back the "This device" button to
dispatch to the platform authenticator when credentials are reported to
be available but not enumerated.

Fixed: 1484660
Change-Id: Ida7b580d1c8d90092d96f8b7c2e0718fa96fd43e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4876021
Reviewed-by: Adam Langley <agl@chromium.org>
Commit-Queue: Nina Satragno <nsatragno@chromium.org>
Auto-Submit: Nina Satragno <nsatragno@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199158}