[Tab Switcher] Fixes for Animate Shrink Expand in Android
Follow ups to:
https://chromium-review.googlesource.com/c/chromium/src/+/4585529

Fix a noticeable x-axis jitter caused by not correctly rounding and
using floats for xOffset in GtsRectAnimator.

Fix an issue with image matrix not being set prior to animation start
by invoking GtsRectAnimator.setRect() with the first rect in the
animation before starting the animation.

Fix an issue where on a non-emulated device, layout might not be
requested again leading to the animation not running. This was
achieved using the same type of logic as TabListRecyclerView does to
indicate it has been laid out prior to the animation starting.

Fix an issue with not calling super for onTabCreated when the flag is
off.

Bug: 1444782
Change-Id: Ibd3e4ac98d50f4fe8d4c33ade29ee64a82e03937
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4875719
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Sky Malice <skym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199170}