[DownloadBubble] Fix empty security subpage
This fixes a bug where the security subpage was shown as empty, and
appeared below the primary view at the same time.

The bug was due to returning to the primary view while the security
view was being shown. This resulted in the primary view and security
view both being shown. We were attempting to return to the primary view
because the danger type was NOT_DANGEROUS. This neglected to take into
account the insecure downloads which don't have a dangerous danger_type.

This CL removes the condition that caused us to return to the primary
view for NOT_DANGEROUS, and just looks at the ui_info instead to see if
it has a subpage.

Bug: 1483517
Change-Id: I37fda6e044466b17076036f3a3c4db3c8beb7389
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4873643
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Commit-Queue: Lily Chen <chlily@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199174}