[no-vary-search-hint] Support waiting for multiple matching prefetches
Add a new NavigationHandleUserData class to store all of the prefetches
the current navigation is blocked waiting for head on.
Then add code for waiting on multiple in-progress prefetches that might
match.
Account for the navigation that gets aborted before we find a prefetch
container to use for the navigation.

Bug: 1462206
Change-Id: Ie99b92e755ac18f3bdcfe428341cb971dbba85d7
Low-Coverage-Reason: COVERAGE_UNDERREPORTED - DVLOG lines are not covered when running tests.
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4808187
Commit-Queue: Liviu Tinta <liviutinta@chromium.org>
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199178}