[ForceSignin] Handle reauth in Profile Picker with a different account
Reauth-ing with a different account that the main one should fail the
flow. Make sure that the different account does not remain signed in
if it was not previously. The flow should end by redirecting the user
to the profile picker with a error dialog message.

Screenshot of the error dialog: https://screenshot.googleplex.com/46JppQ4PmYqgGRR

Bug: 1478217
Change-Id: I42858a9e26a7793e220e1754fdb2efff2a716f2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4874275
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Ryan Sultanem <rsult@google.com>
Cr-Commit-Position: refs/heads/main@{#1199200}