WebView: Moving native initialization to LicenseContentProvider#onCreate
Since WebViewApkApplication.ensureNativeInitialized() is not always
called on the UI thread, we are moving calls in the
LicenseContentProvider to the onCreate method. This will ensure the
native initialization is done on the UIThread when the
LicenseContentProvider is invoked.

Low-Coverage-Reason: This change does not modify how the ContentProvider previously worked. It simply places it on the UIThread.
Bug: 1468203
Change-Id: Icf425e466cebc90c30c97d7889650a0dbdcc373b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4879363
Commit-Queue: Adam Walls <avvall@chromium.org>
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199234}