[SBv5] Log first SafeBrowsing API call response time.
Similar to how the first SafetyNet API call time is logged, add a
histogram Startup.Android.Cold.FirstSafeBrowsingApiResponseTime.Tabbed
to log the first response time from SafeBrowsing API. This will allow
us to compare the speed of first call time between these two APIs.

The name of the old histogram is not changed in this CL for data
continuity. It is a bit confusing because the response is from
SafetyNet API, but the name is FirstSafeBrowsingResponseTime. The
histogram description has been updated to clarify that.

Low-Coverage-Reason: Java code coverage from c++ test not counted
https: //crbug.com/1449041.
Bug: 1444511
Change-Id: Iad979253aa9de82fdf6f785729a8e3bfca91a013
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4870709
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Xinghui Lu <xinghuilu@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Egor Pasko <pasko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199259}