[headless] Handle relative path specification in --disk-cache-dir
net/disk_cache/disk_cache.cc DCHECKs expecting an absolute path
specification when --disk-cache-dir specifies a relative path.

This CL ensures that the specified directory exists and that
an absolute path is used to refer to it.

Drive: since --user-data-dir specification is used to store
disk cache if --disk-cache-dir is not specified, it is also
ensured to have an absolute specification.

Change-Id: I085e9c468786cb70cddcde040371492cb4211763
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4875501
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Peter Kvitek <kvitekp@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199261}