Refactor page visibility state for canvas element
Remove duplicated state in favor of single source of truth:
GetPage()->IsPageVisible().
Rename methods and variables to be clear and consistent: Before this CL
it was not clear that "IsHidden()" referred to page visibility. With
this CL, we use the terms "page visible" in more places and we avoid the
negative form ("hidden") to reduce confusion.

Bug: 1476964
Change-Id: I7d8491814e0cc5f56f1d500fac4c38a35f234685
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4864112
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Justin Novosad <junov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199268}