Prerender: Disallow prerendering for effective URLs
This CL disallows prerendering for effetcive URLs. This usually happens
when hosted apps or remote NTP involve in prerendering.

For hosted apps, this CL handles following cases.

1) Prerendering runs on a hosted app.
2) Prerendering runs for hosted app's URL.
3) Prerendering runs for a URL that redirected to hosted app's URL.
4) Activation runs for hosted app's URL.

Actually, 1) was implemented before this CL. This CL adds test coverage
for the case. 2), 3), and 4) are new cases implemented by this CL. 4) is
a bit tricky: This case happens when prerendering runs for a URL, a
hosted app starts with the URL, and then acivation runs for the URL

For remote NTP, this CL disallows prerendering for the remote NTP URL
that has an effective URL and modifies an existing test.

Change-Id: I904197ece70d52ebecdd078ac3ea5c233a979a80
Bug: 1483169
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4865765
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Lingqi Chi <lingqi@chromium.org>
Reviewed-by: Ken Okada <kenoss@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199269}