[TranslateMessage] Add parameter to show Snackbar
We will test a Snackbar for auto-translated pages when using the
Translate Messages UI. This CL adds a feature parameter to show
the Snackbar and the logic needed to determine if a translation
is automatic or not. A follow up CL will add the Snackbar.

Bug: 1462755
Change-Id: I259722505f1b457df976d5da36236e82af4c0156
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4877321
Commit-Queue: Trevor Perrier <perrier@chromium.org>
Reviewed-by: Basia Zimirska <basiaz@google.com>
Cr-Commit-Position: refs/heads/main@{#1199324}