[X11] Fix scale factor on initialization
This fixes a regression introduced by [1].  The issue is that
OnBeforePlatformScreenInit() is a virtual function that was moved
from Init() to the constructor, so the subclass implementation was
never getting called.

This CL fixes the issue and refactors by removing
DesktopScreenOzoneLinux.  The scale factor plumbing code was written
at a time before dependencies were minimized in //ui/linux:linux_ui.
Now, it's safe for X11 code to depend on //ui/linux:linux_ui directly,
so we can remove all of the plumbing.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/4869739

R=sky
CC=​​oshima,jkummerow

Bug: None
Change-Id: Iab7eedc9c1ee16a1e8c1186a2db015d8d60f4ceb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4879155
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1199345}