Add config_file_version to requests
This change introduces the ability to pass the configuration file version number to the requests made to the server, it is only attached if the flag is enabled. Changing it from attaching a boolean to attaching an integer. Testing the actual response will come in a later CL when the connection with callbacks is made, this is just laying the ground work to pass the value to the requests and to break up the change to make it easier to review.

BUG=b:289117140
TEST=deployed chrome locally and made sure that nothing breaks

Low-Coverage-Reason: test_request_payload is a test helper that doesn't need test coverage

Change-Id: I2c13af107a91f156b2c1e102e09f87a036fc8015
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4876183
Reviewed-by: Hong Xu <xuhong@google.com>
Commit-Queue: Alberto Juarez <albertojuarez@google.com>
Cr-Commit-Position: refs/heads/main@{#1199378}