Fenced frames: Fix ComputeFencedFrameProperties.
Follow up on crrev.com/c/4775457. When ComputeFencedFrameProperties()
is called with `kFencedFrameRoot` and this is an urn navigation to
iframes, the fenced frame properties from the navigation request
should not be returned.

For example, when using this function to find the partition nonce,
it will return the navigation request's fenced frame properties, if it
is not nullopt. Then for urn navigations to iframes, the fenced
frame properties do not have partition nonce, see
`OnFencedFrameURLMappingComplete`. The correct place to obtain the
partition nonce should be from the fenced frame tree root node.

Bug: 1470634
Change-Id: I56974d2ded4c45eaff93c5bca5053b64700b41db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4873916
Commit-Queue: Xiaochen Zhou <xiaochenzh@chromium.org>
Reviewed-by: Garrett Tanzer <gtanzer@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199551}