[ios] Improve WebStateListOrderController & WebStateListRemovingIndexes
Rename WebStateListOrderController and WebStateListRemovingIndexes as
they are (or will soon) become usable independently of WebStateList.

Add an helper structure OrderController::InsertionParams used to pass
the requirements for the DetermineInsertionIndex() method to avoid a
long list of parameters of primitive types (integer and boolean). It
makes the code and the tests simpler to read.

Prepare for making WebStateListOrderController work with other source
(such as a serialized representation of a WebStateList allowing to
optimize SessionRestorationBrowserAgent to not create the empty items
to immediately close them).

Bug: 1383087
Change-Id: I993a1257f62a4eb3a35ef4be4e0ec8b86efd9c1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4874033
Reviewed-by: Asami Doi <asamidoi@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199560}