[bookmarks] Refactor CanBeEditedByUser() and CanSyncNode() to one
These two abstractions are the same in practice and some pieces of code
in fact assume they have the exact same behavior, including an obvious
case reflected in this patch as a removal of a CHECK in
components/browser_sync/sync_client_utils.cc.

Instead, this patch merges the two predicates and adopts a name that
conveys directly the underlying need. The rationale is that, judging
from the existence of clarifying comments in code, it is hard for many
occasional readers/contributors to realize why some bookmarks aren't
editable or syncable, as the concept of enterprise-managed bookmarks
is rather advanced.

Bug: None
Change-Id: Ia276a66aececa285d5e99f91e16f62b9c9130d06
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4875379
Reviewed-by: Brandon Wylie <wylieb@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199605}