[Tracing] Change from base::Uuid to base::Token in trace_report_database
Using base::Token will allow us to easily use the same format as it is
currently used on perfetto. This will lead to be able to set the one
unique ID per trace across the local database and perfetto.

In the following change list:
- Set the trace id through the scenario config.
- Get the trace id from the active scenario config to the pass it to the
NewTraceReport before adding a trace to the database.

Change-Id: I6632001beeb93b01ba52736adaf8a0a1b5b56eba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4864098
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Stephen Nusko <nuskos@chromium.org>
Commit-Queue: Alex Attar <aattar@google.com>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199610}