Expose selection in WebFormControlElement as unsigned
This CL changes the types of the selection indexes in
WebFormControlElement::*Selection*() from `int` to `unsigned`.

The reason is that in HTMLInputElement and HTMLTextAreaElement these
indexes are unsigned anyway, and Autofill (the only caller of the
WebFormControlElement::*Selection* functions) expects/passes unsigned
integers.

Bug: 1007974
Change-Id: I1efbfb70e50dbca1270e1ce6366a06ec89c36b0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4881382
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Commit-Queue: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#1199614}