[iOS] Observe identity update in History Sync screen to update avatar
When signing in with NTP or settings with an account whose avatar was
never cached on the device (new account or avatar changed since last
sign-in), the avatar download can finish after the History Sync
screen's appearance. But the screen didn't observe the corresponding
identity change and wasn't able to update the avatar, so the default
placeholder is always shown.

This CL fixes this by adding an accountManagerServiceObserver to the
HistorySyncMediator, and update the avatar image when an identity
update is received.

Fixed: 1483012
Change-Id: Icc2b39c4f799dab8baa5ae54ec2cd04c3a3d0f33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4882489
Commit-Queue: Menghan Yang <myuu@google.com>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199643}