[MultiInstance] Fix the assertion error
Fixes a corner case where the assert is not valid. When Chrome
Window is switched to another one via menu, with window-task
mapping still present but the corresponding Activity having been
already killed, the assert was triggered, causing confusion.
Now that we know this is a valid case, the assert statement may
be removed.

Bug: 1481331
Change-Id: I1b6ee26ee59c1161481353a74c8282d8ff7507c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4876456
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Reviewed-by: Sirisha Kavuluru <skavuluru@google.com>
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1199673}